requirements.txt
joblib==1.4.2
PyYAML==6.0.2
tqdm==4.66.5
requests>=2.26.0
pandas>=1.3.0
matplotlib>=3.4.0
numpy>=1.20.0
scikit-learn>=1.0.0
seaborn>=0.11.0
torch